add merge_policy::collapse
Ignore-this: 58ace4cb132cd1c78e65a0ca21d7adc5
This is a merge policy for types that should do the sensible thing, in most circumstances.
It avoids conversions between types, unless they are references or pointers to a base and derived class.
It takes the most conservative of the qualifications passed in.

darcs-hash:20141118182153-c9a90-96fb88ce3aaddf269ed19fd1d20915626460122a